Continue restructuring of code, add module `impls`
