+++ src/transition_system/edge.rs
@@ -37,6 +37,46 @@ pub trait IsEdge<'ts, E, Idx, C> {
     }
 }
 
+/// Represents an edge that is not associated to a transition system. It stores a color, an
+/// expression, as well as a source and target state index.
+#[derive(Clone, Eq, PartialEq, Debug)]
+pub struct Edge<E, Idx, C> {
+    source: Idx,
+    target: Idx,
+    color: C,
+    expression: E,
+}
+
+impl<'a, E, Idx: Copy, C: Clone> IsEdge<'a, E, Idx, C> for &'a Edge<E, Idx, C> {
+    fn target(&self) -> Idx {
+        self.target
+    }
+
+    fn color(&self) -> C {
+        self.color.clone()
+    }
+
+    fn expression(&self) -> &'a E {
+        &self.expression
+    }
+
+    fn source(&self) -> Idx {
+        self.source
+    }
+}
+
+impl<E, Idx, C> Edge<E, Idx, C> {
+    /// Creates a new edge with the given source, expression, color and target.
+    pub fn new(source: Idx, expression: E, color: C, target: Idx) -> Self {
+        Self {
+            source,
+            target,
+            color,
+            expression,
+        }
+    }
+}
+
 /// Represents a reference to an edge in a transition system. This stores a lifetime
 /// to the transition system and references to the color and expression.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
