+++ src/transition_system/impls/nts.rs
